fuck me tonight

+++ Suppliers/src/BillOfQuantities.java
@@ -4,17 +4,17 @@ import java.util.Map;
 
 public class BillOfQuantities {
 
-    private Map<Integer, Pair<Integer, Integer>> bill;
+    private Map<Integer, Pair<Integer, Double>> bill;
 
-    public BillOfQuantities(Map<Integer, Pair<Integer, Integer>> bill) {
+    public BillOfQuantities(Map<Integer, Pair<Integer, Double>> bill) {
         this.bill = bill;
     }
 
-    public void insert(Integer itemId, Pair<Integer, Integer> quantity_disc) {
+    public void insert(Integer itemId, Pair<Integer, Double> quantity_disc) {
         bill.put(itemId,quantity_disc);
     }
 
-    public void updateBillOfQuantities(Integer itemId, Pair<Integer, Integer> quantity_disc) {
+    public void updateBillOfQuantities(Integer itemId, Pair<Integer, Double> quantity_disc) {
         bill.put(itemId, quantity_disc);
     }
